fushionbaby-auth/src/main/java/com/fushionbaby/auth/model/AuthPrivilege.java
package com.fushionbaby.auth.model;

/***
 * 
 * @author xupeijun
 * 
 */
public class AuthPrivilege {
	private Long id;

	private String name;

	private Long parentId;

	private Integer sortNo;

	private String url;
	
	private Integer level;

	/** 做页面显示 */
	private boolean select;

	/** 一级权限的图标 */
	private String iconClass;

	public String getIconClass() {
		return iconClass;
	}

	public void setIconClass(String iconClass) {
		this.iconClass = iconClass;
	}

	public Long getId() {
		return id;
	}

	public void setId(Long id) {
		this.id = id;
	}

	public String getName() {
		return name;
	}

	public void setName(String name) {
		this.name = name == null ? null : name.trim();
	}

	public Long getParentId() {
		return parentId;
	}

	public void setParentId(Long parentId) {
		this.parentId = parentId;
	}

	public Integer getSortNo() {
		return sortNo;
	}

	public void setSortNo(Integer sortNo) {
		this.sortNo = sortNo;
	}

	public String getUrl() {
		return url;
	}

	public void setUrl(String url) {
		this.url = url == null ? null : url.trim();
	}

	public Integer getLevel() {
		return level;
	}

	public void setLevel(Integer level) {
		this.level = level;
	}

	public boolean isSelect() {
		return select;
	}

	public void setSelect(boolean select) {
		this.select = select;
	}
}